requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.32
pydantic==2.9.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
